[Packager] Rename `_read` to `read` in Module, Packager and AssetModule classes.
Renaming `_read` method in Module.js and Packager.js to `read`, as to not be a private method. This is done in order to allow for proper class inheritance, even when using Facebook's internal "babel-plugin-class-private-props" transform, which actually mangles method names starting with '_', preventing child classes from overriding the method.