Adds proper layout measurement for the ReactTextViewManager.
In order to leverage the CSSNode class to its fullest, various view managers need to override the measurement algorithm to determine the proper width and height needed for a view.

For now, we needed to move the measurement of FrameworkElements to the dispatcher thread, as WinRT does not have any text formatter or other hooks for determining the size of text without mimicking the element itself.  Issue #106 covers this.