Use "babel-preset-react-native"
Rather than specifying Babel plugins in the `.babelrc` packaged with
react-native, leverage a Babel preset to define the plugins
(https://github.com/exponentjs/babel-preset-react-native).

This allows for a much better user experience for those who want (or
need) to override options in their project's `.babelrc`.

Prior to this PR, if a user wanted to use a custom babel-plugin (or a
custom set of babel plugins), they'd have either 1) manually override
the `.babelrc` in the react-packager directory (or fork RN), or 2)
specify a custom transformer to use when running the packager that
loaded their own `.babelrc`. Note - the custom transformer was
necessary because without it, RN's `.babelrc` options would supersede
the options defined in the project's `.babelrc`...potentially causing
issues with plugin ordering.

This PR makes the transformer check for the existence of a
project-level `.babelrc`, and if it it's there, it _doesn't_ use the
react-native `.babelrc`. This prevents any oddities with Babel plugin
ordering, and leaves that up to the RN user. The RN user would then
just include a `.babelrc` file in their project that looks like:

```
{
"presets": ["react-native"]
}
```

They can then add whatever other presets or plugins that they'd like
(such as the Babel Relay Plugin or a decorators plugin, for example).