Make the packager work with babel strict mode transform
At the moment we have to disable strict mode for the
transform-es2015-modules-commonjs because strict mode leaks to the global
scope and breaks the bridge. It was due to the way the polyfills were bundled in the package. To fix it, I wrapped the polyfill modules in an IIFE. Then when strict mode was enabled some polyfills were broken due to strict mode errors so that was fixed too. Also removed the IIFE from the polyfills that included one.