Add RCTJavaScriptContextCreatedNotification for iOS
This notification will be fired on the JS thread after a JSContext has been created and setup, but before the JS bundle is loaded. This allows for advanced native modules to add custom functionality (that isn't possible with the existing APIs) into the context that will be available when JS bundle loads.