Fixes issue with dispose behavior on MessageQueueThread instances.
Previously, there were no guarantees on the order of operations between items that may have been enqueued prior to disposal and the disposal itself.

In order for unit tests to behave properly, background threads are not allowed to crash. These changes prevent background threads from crashing by imposing an order of operations on the disposal of message queue threads.