Add a deprecatedPropType module to show deprecation warnings
To allow smoother API changes for users we often deprecate props and keep
them around for a while before removing them. Right now it is all done
manually, this adds a consistent way to show a warning when using a
deprecated prop.

It also changes places where we added the warnings manually to use
deprecatedPropType instead.