Fix bug related to removeCLippedSubviews and view collapsing.
The bug occurs for the cases when there is a nested view structure of at least two views with removeClippedSubvews enabled and with a "collapsable" view in between them that migrates from the collapsed state to non-collapsed state.

What happens in that case is that the "inner" view with "removeClippsedSubviews" gets reattached to a new parent, but we never update it's clipping rect because the update is currently only triggered for the size change (and for scroll change in case of scrollview). In the case when the view was doing some "clipping" when attached to its previous parent it needs to update its "clipping" status because the parent has change and clipping rect is calculated based on the parent clipping rect (see `ReactClippingViewGroupHelper#calculateClippingRect`).

This change triggers `updateClippingRect` when the view is attached to the window, which covers the case when it's reattached from one parent to the other.

Test plan: Here is a simple app that illustrates the issue: https://gist.github.com/kmagiera/f2af6d6b9217fcacfea6
Without this change the "black" square won't be visible after tapping "yellow" button twice
Although it will appear when we turn off "clipping" or "collapsing", so for the following cases:
REMOVE_CLIPPED=false or REMOVE_CLIPPED=true