Check ListView SCROLLVIEW_REF can scroll
If rendering a ListView with a component that doesn't scroll - for example a standard View - the getScrollResponder method should return false. Otherwise a "is not a function" exception is thrown.

This was slightly addressed https://github.com/facebook/react-native/commit/4326b01181559a908740cda0a4c8459c9bdb919d but not quite fully.