Fixes #191 - Change JSON marshaling to use parse and stringify
After running throughput performance experiments, it was discovered that it's actually faster to marshal JSON data by first stringifying in C#, sending as a string value to Chakra, and then parsing in Chakra (and the inverse applies coming from Chakra to C#).

This changeset swaps out the visitor pattern for generating Chakra JavaScript values for the serialize / parse method.