Update Navigator.js
I've had consistently seen questions pop up on StackOverflow about passing properties using Navigator, so I thought it may help to give an example of how to do this. I've heard the navigator api may be changing or being replaces soon, but until then there seems to be a need for this info somewhere. Maybe this is not the best way to present it, but hopefully it will be a start. Thanks!