Updating the `react-native run-ios` command to use UDID
Updating the `react-native run-ios` command to launch the simulator based off the UUID and to retrieve the list of simulators with the --json flag and therefor removing the reliance on regex parsing and the `local-cli/runIOS/parseIOSSimulatorsList.js`.

This solves a few problems:
1) 	When trying to launch an iOS simulator with just the name when you have other simulators that are similarly named.
	For example, you may have a simulator called 'iPhone 6' and one called iPhone 6 + 38mm Apple Watch'.
	If you try and run `xcrun instruments -w "iPhone 6"` you will get an error Instruments Usage Error : Ambiguous device name/identifier 'iPhone 6'.

2)	If there is a simulator with brackets in it's name (e.g. "iPhone 6 (test 2)") the regex will parse the contents of the contents of the brackets as the UDID.
	This isn't an issue prior to this change for launching the simulator, but the UDID is used for building the project so it would have failed there.

3)	It won't launch unavailable simulators. Because we can accurately get the status of each simulator, we can filter out ones that are not in the available state.
	This could be extended to check if they are already booted, or to use the default one as the booted one.

All in all, this feels like a must more robust way of parsing the output of the list of simulators.

I've also removed the default of 'iPhone 6' for the currently booted simulator, or the first in the list. If there is a simulator that is booted, we can't boot any others.