Allows a user to specify an xcode scheme name.
When running react-native run-ios, this feature allows a user to specify which scheme to run.

Defaults to previous behavior of inferring the scheme name based on the project file name  if the parameter is not present.

Parameter is not required and running react-native run-ios --help shows the parameter's description.

Example usage:
react-native run-ios --scheme myDifferentScheme