Remove the need for allowTopLevelThis in transform-es2015-modules-commonjs
This removes the few places that a top level this was used to refer to
the global space. It also clean up the usage of `GLOBAL` to use `global`
instead as this is what is used everywhere else in the code base. We still
define `GLOBAL` for compatibility with other modules.