[Tests] Add a test for RCTBridgeModules that aren't auto-exported
Adds a test so that people can create RCTBridgeModules that aren't auto-exported. This is useful for when you have more than one RCTBridge and want a module to be exposed in one but not the other.

RCTBridge/RCTBatchedBridge already supports this functionality; this diff adds a test so that it doesn't break.

Test Plan: Run the iOS UIExplorer unit tests from Xcode.