[ListView] Use function refs and support composed refs
Fixes an issue where if you implement `renderScrollComponent` and have a `ref` callback on the returned element, the ref is not clobbered by the ref that ListView adds to the the element.

This is accomplished by converting the ref from a legacy string-based ref to a callback-based ref, and then using `cloneReferencedElement`, which is a simple utility to compose callback refs.

Test Plan: Open the ListView examples in UIExplorer