Fixed TouchableNativeFeedback.Ripple to follow subview borderRadii
Non-ripples generated by TNF.Ripple are now clipped by border radii the
same way that the ReactViewGroup it's a background of gets clipped. This
is achieved by refreshing those backgrounds in ReactViewGroup whenever
nativeBackground or any of the borderRadii is updated, rather than
ReactViewManager just setting the background Drawable. Note that
TNF.SelectableBackground is not clipped because it's not possible to
clip the selectableItemBackground Drawable.