Upgrade ndk-build to use gcc 4.9
Recently released ndk-11b does no longer have gcc 4.8. In addition google no longer provide links for older NDK version downloads (one need to copy a link address to the most recent version and swap version part in the URL, which is the only way to download that).

In order for the native code to work after upgrading to 4.9 I had to update c++ flags so that the sources are now compiled in c++1y standard - this is required as 4.9 doesn't allow 'lambda capture initializers' (added recently to OnLoad.cpp and JSCExecutor.cpp).