[init] Honor the version of "react" under peerDeps when setting up a project
We need this since React 15.0.0 is coming and will break `react-native init`, which currently installs the latest version of React. We'll need some changes to React Native to support 15 that Sebastian is actively working on, but till that lands we want `react-native init` to continue working.

Test Plan: Run `scripts/test-manual-e2e.sh` and verify that it ran `npm install react@^0.14.5 --save` instead of `npm install react --save`.