[Animated] Use native animations in TouchableBounce and TouchableOpacity
Now that native animations for opacity and springs have landed in both iOS and Android, we can enable native animations both for TouchableBounce and TouchableOpacity.

Test Plan: Open the UIExplorer's Touchable demo and tap the "Enabled TouchableOpacity" component. Then change it to be TouchableBounce and reload the app and verify that works too. Do this on iOS and Android.