Fix system font issue
The problem here was - when no fontFamily was specified, we were not
really using `System` (in fact we should since that's default).

Also, assuming I have passed `System` as a fontFamily and we were
returning correct UIFont, next time I was passing that font to
RCTConvert font:previouslyMadeFont withWeight:@"bold" - it was no longer
taking it as a system font since the fontFamily was now .SF UI Text - it
was falling under non-system case.

Note: The non-system case is still broken as I haven't got to it yet.