StatusBar imperative API: honor animation defaults for optionals
StatusBar's imperative API contains optional animation inputs in the
methods `setHidden`, `setBarStyle` and `setBackgroundColor`, but not
passing values currently defaults to no animation, even when a
StatusBar instance at the top of the app declares animation props.

Since these animation inputs are declared as optional and no method
for setting up global defaults is provided, from a developer
perspective, it makes sense to honor the current props stack’s
animation settings.

For example if this is included at the top of the app,

`<StatusBar animated={true} showHideTransition='slide' />`

it intuitively makes most sense that calling
`StatusBar.setHidden(true)`
would still show a slide animation, instead of needing to pass
`'slide'`.

A slightly more architecture-correct way to fix this would be to add a
`StatusBar.setDefaultProps(props:Object)` method, but that would
require changing the API and docs. The current docs seem to hint that
it is possible to set defaults on a top-level StatusBar instance, so it
feels simplest to just honor those animation defaults when defined.