iOS: Support HTTP headers for source prop on <Image> components
Allows developers to specify headers to include in the HTTP request
when fetching a remote image. For example, one might leverage this
when fetching an image from an endpoint that requires authentication:

```
<Image
  style={styles.logo}
  source={{
    uri: 'http://facebook.github.io/react/img/logo_og.png',
    headers: {
      Authorization: 'someAuthToken'
    }
  }}
/>
```

Note that the header values must be strings.

When doing image requests, this also gives developers control over the
other properties supported by [RCTConvert NSURLRequest:] like `method`
and `body`.

This change deprecates some APIs on RCTImageLoader in order to replace
them with more flexible APIs which take an NSURLRequest rather than just
an NSURL.