ScrollView: Always fire onScroll event for the resting scroll position
When throttling scroll events with scrollEventThrottle, onScroll
is not guaranteed to be fired for the final scroll position
of the ScrollView. This can cause a component to render UI that
is consistent with the resting scroll position of the ScrollView.

This commit guarantees that an onScroll event will be fired for
the resting scroll position of the ScrollView.