NavigatorIOS: Expose interactivePopGestureEnabled property
Previously, the back swipe navigation gesture would be enabled when
the navigation bar is shown and disabled when the navigation bar
is hidden.

This change enables developers to control the back swipe gesture
independently of the visibility of the navigation bar. An example
use case would be that an app wants to render a custom navigation
bar so it sets navigationBarHidden to true and it wants to enable
the back swipe gesture so it sets interactivePopGestureEnabled to
true.