Added additional predefined easing options to Easings.js
Mapped the first 24 predefined Easing curves on http://easings.net/ for quick use. There are 6 more easing curves on that website, however they don't supply a bezier curve for them so we should probably leave those up to developers. Usually I would wrap all these in an object called easingTypes or something and set them to be const however bases on how the first one was set I decided to follow the standard. Let me know if I should wrap/export these cleaner. Also I would like to update the Animated docs page with this information, would you like me to do that in this PR? Or is that something that you guys usually handle.