Create Core Components Tutorial (1/N)
This commit create the skeleton for our Tutorials section.

We are going to start with core component Tutorials (e.g. `View).

Each of the samples associated with the core components will
(hopefully) be runnable in our online simulator when we add that.

This also reorganizes the documentation a bit as well. For example,
what was our tutorial is now under our Sample Apps section.