Make the root view background color explicit
The default white background for the root view causes an unattractive flash when loading any app with a non-white background.  Almost every developer has to search for how to fix it.  

This change makes the background color explicit in the file, so it can be easily change to match the color of the splash screen and the main screen without having to search for it.