Hook up Android intent to Linking.js #7079
This PR references the deep linking issue in #7079. `IntentAndroid`
states that it is deprecated in favor of Linking, but Linking [states
that `addEventListener` and `removeEventListener` are not supported in
Android](https://github.com/facebook/react-native/blob/master/Libraries/Linking/Linking.js#L122).
This PR adds both methods by implementing the `onNewIntent` method for
ReactActivity and passing the result through DeviceEventEmitter to
Linking.js.