Adds the ability to use UIManager to check if a node is an ancestor of other node.
Updated tests to latest version

Adds the ability to use UIManager to check if a node is an ancestor of other node.

Updated tests to latest version

Adds the ability to use UIManager to check if a node is an ancestor of other node.

Added @alloy suggestions and refactored the descendant method name