Fix bug where library is really written to the current project's node_modules
Currently the library is dumped at `SomeProject/node_modules/react-native/Libraries/MyLibrary` because the target path is being build from the sample path, since the code takes the template (`Sample`) from that path. However the code mistakenly assumes this is also the destination path.

The effect is that the actual library been generated sits at the project's `node_modules` which is not under source control, and the console mistakenly reports that the library sits within the project `Libraries` folder, by using the correct `libraryDest`.

This fix trims the internal path that `Sample` has so that the files are actually placed in the project's source tree and not within `node_modules`.