Move Component Embedded Simulator next to its example
Right now the embedded simulator is always at the top right corner.
This can be confusing as to what code is associated with the simulation.

So, move the simulator next to its actual code.

This has the added benefit of allowing us to use the React Native
Web Player for the simpler examples in the components.

Test Plan:

http://localhost:8079/react-native/docs/switch.html