More information for valid inputs
I was having difficultly understanding what values were valid for the shadow. The documentation was a little sparse. I jumped on rnplay and tested some numbers and these are the descriptions which I came up with based on my experimentation I would love for these or something even more accurate to make it into the documentation.